ai-pipeline/src/models/chest_xray_detector.py
"""
Chest X-ray anomaly detection model using deep learning.
This is a simplified implementation for demonstration purposes.
"""

import cv2
import numpy as np
from typing import List, Dict, Tuple
import torch
import torch.nn as nn
from torchvision import transforms
from PIL import Image


class ChestXrayDetector:
    """
    AI model for detecting anomalies in chest X-ray images.
    """
    
    def __init__(self, model_path: str = None):
        self.model_path = model_path
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model = self._load_model()
        self.transform = self._get_transforms()
        
        # Predefined detection classes
        self.classes = [
            'Pulmonary Nodule',
            'Consolidation',
            'Calcification',
            'Pleural Effusion',
            'Pneumothorax',
            'Cardiomegaly',
            'Atelectasis',
            'Infiltration'
        ]
    
    def _load_model(self):
        """Load the trained model. In production, this would load actual weights."""
        # Simplified model architecture for demonstration
        model = nn.Sequential(
            nn.Conv2d(1, 32, 3, padding=1),
            nn.ReLU(),
            nn.MaxPool2d(2),
            nn.Conv2d(32, 64, 3, padding=1),
            nn.ReLU(),
            nn.MaxPool2d(2),
            nn.AdaptiveAvgPool2d((1, 1)),
            nn.Flatten(),
            nn.Linear(64, len(self.classes))
        )
        
        if self.model_path:
            # In production: model.load_state_dict(torch.load(self.model_path))
            pass
            
        model.to(self.device)
        model.eval()
        return model
    
    def _get_transforms(self):
        """Get image preprocessing transforms."""
        return transforms.Compose([
            transforms.Grayscale(),
            transforms.Resize((512, 512)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485], std=[0.229])
        ])
    
    def preprocess_image(self, image_path: str) -> np.ndarray:
        """
        Preprocess medical image for analysis.
        
        Args:
            image_path: Path to the medical image
            
        Returns:
            Preprocessed image array
        """
        # Load image
        image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
        
        if image is None:
            raise ValueError(f"Could not load image from {image_path}")
        
        # Apply medical image preprocessing
        # Contrast enhancement
        clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8, 8))
        image = clahe.apply(image)
        
        # Noise reduction
        image = cv2.bilateralFilter(image, 9, 75, 75)
        
        # Normalize
        image = cv2.normalize(image, None, 0, 255, cv2.NORM_MINMAX)
        
        return image
    
    def detect_anomalies(self, image_path: str) -> List[Dict]:
        """
        Detect anomalies in chest X-ray image.
        
        Args:
            image_path: Path to the chest X-ray image
            
        Returns:
            List of detected anomalies with bounding boxes and confidence scores
        """
        # Preprocess image
        processed_image = self.preprocess_image(image_path)
        
        # For demonstration, we'll simulate AI detections
        # In production, this would use the actual trained model
        detections = self._simulate_detections(processed_image)
        
        return detections
    
    def _simulate_detections(self, image: np.ndarray) -> List[Dict]:
        """
        Simulate AI detections for demonstration purposes.
        In production, this would be replaced with actual model inference.
        """
        height, width = image.shape
        
        # Simulate realistic detections
        detections = [
            {
                'id': 'ai-1',
                'label': 'Pulmonary Nodule',
                'confidence': 0.89,
                'bounding_box': {
                    'x': int(width * 0.4),
                    'y': int(height * 0.3),
                    'width': int(width * 0.056),
                    'height': int(height * 0.063)
                },
                'description': 'Small rounded opacity in right upper lobe',
                'severity': 'medium'
            },
            {
                'id': 'ai-2',
                'label': 'Consolidation',
                'confidence': 0.76,
                'bounding_box': {
                    'x': int(width * 0.6),
                    'y': int(height * 0.47),
                    'width': int(width * 0.085),
                    'height': int(height * 0.087)
                },
                'description': 'Area of increased opacity suggesting consolidation',
                'severity': 'high'
            },
            {
                'id': 'ai-3',
                'label': 'Calcification',
                'confidence': 0.92,
                'bounding_box': {
                    'x': int(width * 0.475),
                    'y': int(height * 0.37),
                    'width': int(width * 0.015),
                    'height': int(height * 0.025)
                },
                'description': 'Small calcified granuloma',
                'severity': 'low'
            }
        ]
        
        return detections
    
    def extract_measurements(self, image_path: str, pixel_spacing: float = 0.14) -> Dict:
        """
        Extract measurements from detected anomalies.
        
        Args:
            image_path: Path to the image
            pixel_spacing: Pixel spacing in mm/pixel
            
        Returns:
            Dictionary containing measurements
        """
        detections = self.detect_anomalies(image_path)
        measurements = {}
        
        for detection in detections:
            bbox = detection['bounding_box']
            width_mm = bbox['width'] * pixel_spacing
            height_mm = bbox['height'] * pixel_spacing
            area_mm2 = width_mm * height_mm
            
            measurements[detection['id']] = {
                'width_mm': round(width_mm, 2),
                'height_mm': round(height_mm, 2),
                'area_mm2': round(area_mm2, 2),
                'label': detection['label']
            }
        
        return measurements