ai-pipeline/requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
scipy==1.11.4

# Machine Learning
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
scikit-image==0.22.0

# Medical imaging
pydicom==2.4.3
nibabel==5.1.0
SimpleITK==2.3.1

# AWS services
boto3==1.34.0
botocore==1.34.0

# API and web services
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0

# Image processing
imageio==2.33.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1